Added view for upcoming appoinments on employee end, updated appearing appointments
View for upcoming appointments: Employees can look at appointments they must conduct in a table, with the patient name, their ID, room number, start/end time, and date of appointment.

Appointments are limited by date, if the date of the appointment is the current date or later then it will display.

BUG: Time displays wrong on both employee.jsp and viewUpAppointments.jsp.