Added edit employee profile, added employee list to patient view, renamed getColumns for queries, changed session attribute from email to userid
Edit Employee Profile: Only edits address, phone, gender, email, and password.

Employee List for Patient: Displays table of employees that are assigned to the patient.

Changed resultset gets with the column names. Before it was the number of the column. Changed for readability.

Changed from email to userid. email constantly changes, could potentially mess up when editing personal information. User id is constant and doesn't change. Receives the user id from the login page and continually uses it for the rest of the session.